# model/web/datakinerjaopd_update_request.go
package web

type DataKinerjaOpdUpdateRequest struct {
	Id                   int                   `validate:"required" json:"id"`
	JenisDataId          int                   `validate:"required" json:"jenis_data_id"`
	KodeOpd              string                `validate:"required" json:"kode_opd"`
	NamaData             string                `validate:"required" json:"nama_data"`
	RumusPerhitungan     string                `validate:"required" json:"rumus_perhitungan"`
	SumberData           string                `validate:"required" json:"sumber_data"`
	InstansiProdusenData string                `validate:"required" json:"instansi_produsen_data"`
	Target               []TargetUpdateRequest `validate:"required" json:"target"`
	Keterangan           string                `json:"keterangan"`
}